Move generate_kfrags() to the pre module
