+++ umbral/pre.py
@@ -1,10 +1,11 @@
-from typing import Tuple, Optional, Sequence
+from typing import Tuple, Optional, Sequence, List
 
 from .capsule import Capsule
 from .capsule_frag import VerifiedCapsuleFrag, CapsuleFrag
 from .dem import DEM
 from .keys import PublicKey, SecretKey
-from .key_frag import VerifiedKeyFrag, KeyFrag
+from .key_frag import VerifiedKeyFrag, KeyFrag, KeyFragBase
+from .signing import Signer
 
 
 def encrypt(pk: PublicKey, plaintext: bytes) -> Tuple[Capsule, bytes]:
@@ -28,6 +29,34 @@ def decrypt_original(sk: SecretKey, capsule: Capsule, ciphertext: bytes) -> byte
     key_seed = capsule.open_original(sk)
     dem = DEM(bytes(key_seed))
     return dem.decrypt(ciphertext, authenticated_data=bytes(capsule))
+
+
+def generate_kfrags(delegating_sk: SecretKey,
+                    receiving_pk: PublicKey,
+                    signer: Signer,
+                    threshold: int,
+                    num_kfrags: int,
+                    sign_delegating_key: bool = True,
+                    sign_receiving_key: bool = True,
+                    ) -> List[VerifiedKeyFrag]:
+    """
+    Generates ``num_kfrags`` key fragments to pass to proxies for re-encryption.
+    At least ``threshold`` of them will be needed for decryption.
+    If ``sign_delegating_key`` or ``sign_receiving_key`` are ``True``,
+    the corresponding keys will have to be provided to :py:meth:`KeyFrag.verify`.
+    """
+
+    base = KeyFragBase(delegating_sk, receiving_pk, signer, threshold)
+
+    # Technically we could allow it, but what would be the use of these kfrags?
+    if num_kfrags < threshold:
+        raise ValueError(f"Creating less kfrags ({num_kfrags}) than threshold ({threshold}) makes them useless")
+
+    kfrags = [KeyFrag.from_base(base, sign_delegating_key, sign_receiving_key)
+              for _ in range(num_kfrags)]
+
+    # Make them verified - we know they're good.
+    return [VerifiedKeyFrag(kfrag) for kfrag in kfrags]
 
 
 def reencrypt(capsule: Capsule,
